# back-end/swagger.json
{
    "swagger": "3.0",
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "AgileOne CRUD",
        "description": "Recruiting Application",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://localhost:2000/api/",
            "description": "Dev"
        }
    ],
    "components": {        
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "schemes": ["http","https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/admin/get_employees": {
            "post": {
                "tags": ["Recruiters"],
                "description": "Get Recruiters",
                "operationId": "get_employees",
                "summary": "Get Recruiters Details in system",
                "consumes": "application/json",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "perPage": "10/20/30",
                                "page": "Page Number",
                                "search": "Search Text",
                                "role": "Role ID from get_roles api",
                                "team_leads": "Pass Teams Lead Ids as an array when user login as Accounts Manager",
                                "team_lead_id": "Team Lead ID"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/get_allemployees": {
            "post": {
                "tags": ["Recruiters"],
                "description": "Get All Recruiters",
                "operationId": "get_allemployees",
                "summary": "Get All Recruiters Details in system",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "team_lead_id": "Team Lead ID"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/create_employee": {
            "post": {
                "tags": ["Recruiters"],
                "description": "Add Recruiter Details",
                "operationId": "create_employee",
                "summary": "Add Recruiter Details in system",
                "consumes": "application/json",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "employee_name": "Recruiter Name",
                                "employee_id": "Recruiter ID",
                                "mobile_number": "Mobile Number",
                                "password": "Recruiter Password",
                                "email": "Recruiter Email",
                                "office_email": "Recruiter Office Email",
                                "address": "Recruiter Address",
                                "designation": "Recruiter Designation",
                                "team_lead_id": "Team Lead ID"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/update_employee": {
            "post": {
                "tags": ["Recruiters"],
                "description": "Update Recruiter Details",
                "operationId": "update_employee",
                "summary": "Update Recruiter Details in system",
                "consumes": "application/json",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "employee_name": "Recruiter Name",
                                "employee_id": "Recruiter ID",
                                "mobile_number": "Mobile Number",
                                "email": "Recruiter Email",
                                "office_email": "Recruiter Office Email",
                                "address": "Recruiter Address",
                                "designation": "Recruiter Designation",
                                "id": "Recruiter ID"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/get_singleemployee": {
            "post": {
                "tags": ["Recruiters"],
                "description": "Get Single Recruiter Details",
                "operationId": "get_singleemployee",
                "summary": "Get Single Recruiter Details in system",
                "consumes": "application/json",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "user_id": "Recruiter ID"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
}